# Conflicts:
#	cypress/e2e/app-preferences.cy.ts
#	cypress/e2e/application-project-setting.cy.ts
#	cypress/e2e/clone-map-tests.cy.ts
#	cypress/e2e/config-template.cy.ts
#	cypress/e2e/device-template.cy.ts
#	cypress/e2e/images-icon.cy.ts
#	cypress/e2e/ip-subnet.cy.ts
#	cypress/e2e/login-profile.cy.ts
#	cypress/e2e/map-basic-features.cy.ts
#	cypress/e2e/map-preferences.cy.ts
#	cypress/e2e/node-view.cy.ts
#	cypress/e2e/project-import-export.cy.ts
#	cypress/fixtures/map/map-features.json
#	cypress/support/commands.ts
#	src/app/administration/roles/roles.component.ts
#	src/app/app.component.ts
#	src/app/core/enums/api-paths.enum.ts
#	src/app/core/interceptors/auth/auth.interceptor.ts
#	src/app/core/main-layout/nav-bar/nav-bar.component.html
#	src/app/core/main-layout/nav-bar/nav-bar.component.ts
#	src/app/core/services/auth/auth.service.ts
#	src/app/core/services/config-template/config-template.service.ts
#	src/app/core/services/device/device.service.ts
#	src/app/core/services/hardware/hardware.service.ts
#	src/app/core/services/helpers/helpers.service.ts
#	src/app/core/services/map-edit/map-edit.service.ts
#	src/app/core/services/roles/roles.service.ts
#	src/app/core/services/server-connect/server-connect.service.ts
#	src/app/core/services/template/template.service.ts
#	src/app/core/storage/local-storage/local-storage-data.model.ts
#	src/app/core/storage/local-storage/local-storage-keys.enum.ts
#	src/app/devices/hardware/hardware.component.ts
#	src/app/devices/login-profiles/login-profiles.component.html
#	src/app/libraries/config-templates/add-edit-config-template/add-edit-config-template.component.html
#	src/app/libraries/config-templates/add-edit-config-template/add-edit-config-template.component.ts
#	src/app/libraries/config-templates/config-templates.component.html
#	src/app/libraries/config-templates/config-templates.component.ts
#	src/app/libraries/image/icon-gallery/icon-gallery.component.html
#	src/app/libraries/image/images/images.component.html
#	src/app/login/login.component.ts
#	src/app/map/add-node-pg-to-groupbox-dialog/add-node-pg-to-groupbox-dialog.component.ts
#	src/app/map/add-update-group-dialog/add-update-group-dialog.component.ts
#	src/app/map/add-update-interface-dialog/add-update-interface-dialog.component.html
#	src/app/map/add-update-interface-dialog/add-update-interface-dialog.component.scss
#	src/app/map/add-update-interface-dialog/add-update-interface-dialog.component.ts
#	src/app/map/add-update-node-dialog/add-update-node-dialog.component.html
#	src/app/map/add-update-node-dialog/add-update-node-dialog.component.scss
#	src/app/map/add-update-node-dialog/add-update-node-dialog.component.ts
#	src/app/map/add-update-pg-dialog/add-update-pg-dialog.component.html
#	src/app/map/add-update-pg-dialog/add-update-pg-dialog.component.ts
#	src/app/map/context-menu/cm-actions/cm-actions.service.ts
#	src/app/map/context-menu/cm-group-option/cm-group-option.service.ts
#	src/app/map/context-menu/context-menu.service.ts
#	src/app/map/info-panel/info-panel-interface-management/info-panel-interface-management.component.ts
#	src/app/map/info-panel/info-panel-node/info-panel-node.component.ts
#	src/app/map/info-panel/info-panel-port-group-management/info-panel-port-group-management.component.ts
#	src/app/map/map.component.html
#	src/app/map/map.component.ts
#	src/app/map/map.module.ts
#	src/app/map/tool-panel/tool-panel-edit/tool-panel-edit.component.html
#	src/app/map/tool-panel/tool-panel-edit/tool-panel-edit.component.ts
#	src/app/map/tool-panel/tool-panel-option/tool-panel-option.component.html
#	src/app/map/tool-panel/tool-panel-style/tool-panel-style.component.html
#	src/app/map/tool-panel/tool-panel.component.html
#	src/app/map/tool-panel/tool-panel.component.ts
#	src/app/project/add-project/add-project.component.ts
#	src/app/project/edit-project-dialog/edit-project-dialog.component.ts
#	src/app/project/project.component.ts
#	src/app/project/services/project.service.ts
#	src/app/remote/connection-profiles/connection-profiles.component.html
#	src/app/remote/connection-profiles/connection-profiles.component.ts
#	src/app/settings/map-preferences/map-preferences.component.ts
#	src/app/shared/components/import-dialog/import-dialog.component.html
#	src/app/shared/components/import-dialog/import-dialog.component.ts
#	src/app/shared/enums/error-messages.enum.ts
#	src/app/shared/validations/networks.validation.ts

+++ src/app/map/context-menu/cm-map/cm-map.service.ts
@@ -1,8 +1,9 @@
 import { Injectable } from '@angular/core';
 import { Store } from '@ngrx/store';
+import { ToastrService } from 'ngx-toastr';
 import { HelpersService } from 'src/app/core/services/helpers/helpers.service';
 import { retrievedMapContextMenu } from 'src/app/store/map-context-menu/map-context-menu.actions';
-import { retrievedMapSelection } from 'src/app/store/map-selection/map-selection.actions';
+import { selectAllElementsOnMap, unSelectAllElementsOnMap } from "../../../store/map/map.actions";
 
 @Injectable({
   providedIn: 'root'
@@ -12,6 +13,7 @@ export class CMMapService {
   constructor(
     private store: Store,
     private helpersService: HelpersService,
+    private toastr: ToastrService,
   ) { }
 
   getSaveChangesMenu(isCanWriteOnProject: boolean) {
@@ -72,19 +74,21 @@ export class CMMapService {
       content: "Lock All",
       coreAsWell: true,
       onClickFunction: (event: any) => {
-        cy.nodes().forEach((ele: any) => {
-          const d = ele.data();
-          if (d.elem_category != 'group') {
-            ele.lock();
-            d.locked = true;
-            if (!d.new) {
-                d.new = false;
-                d.updated = true;
-                d.deleted = false;
+        const lockedNodes = cy.nodes().filter((node: any) => node.data().locked)
+        if (lockedNodes.length === cy.nodes().length) {
+          this.toastr.warning("Already Locked");
+        } else {
+          cy.nodes().forEach((ele: any) => {
+            const d = ele.data();
+            if (d.elem_category != 'group') {
+              ele.lock();
+              d.locked = true;
+              d.updated = true;
+              this.helpersService.addBadge(cy, ele);
             }
-            this.helpersService.addBadge(cy, ele);
-          }
-        });
+          });
+          this.toastr.success("Locked all elements");
+        }
       },
       hasTrailingDivider: false,
       disabled: false,
@@ -97,69 +101,45 @@ export class CMMapService {
       content: "Unlock All",
       coreAsWell: true,
       onClickFunction: (event: any) => {
-        cy.nodes().forEach((ele: any) => {
-          const d = ele.data();
-          if (d.elem_category != 'group') {
-            ele.unlock();
-            d.locked = false;
-            if (!d.new) {
-                d.new = false;
-                d.updated = true;
-                d.deleted = false;
+        const unLockedNodes = cy.nodes().filter((node: any) => !node.data().locked)
+        if (unLockedNodes.length === cy.nodes().length) {
+          this.toastr.warning("Already Unlocked");
+        } else {
+          cy.nodes().forEach((ele: any) => {
+            const d = ele.data();
+            if (d.elem_category != 'group') {
+              ele.unlock();
+              d.locked = false;
+              d.updated = true;
+              this.helpersService.removeBadge(ele);
             }
-            this.helpersService.removeBadge(ele);
-          }
-        });
+          });
+          this.toastr.success("Unlocked all elements");
+        }
       },
       hasTrailingDivider: false,
       disabled: false,
     }
   }
 
-  getSelectAllMenu(cy: any, activeNodes: any[], activePGs: any[], activeEdges: any[], activeGBs: any[], activeMBs: any[], activeMapLinks: any[]) {
+  getSelectAllMenu(cy: any) {
     return {
       id: "select_all",
       content: "Select All",
       coreAsWell: true,
       onClickFunction: (event: any) => {
-        const allNodes = cy.nodes();
-        const allEdges = cy.edges();
         cy.nodes().selectify();
         cy.edges().selectify();
-        const activeEles = activeNodes.concat(activePGs, activeEdges);
-        if (activeEles.length == 0) {
-          activeNodes.splice(0);
-          activePGs.splice(0);
-          activeEdges.splice(0);
-          activeGBs.splice(0);
-        }
-        const allNodeAndEdges = allNodes.merge(allEdges)
-        allNodeAndEdges.forEach((node: any) => {
-          const d = node.data();
-          if (d.elem_category == 'node' && !activeNodes.includes(node)) {
-            activeNodes.push(node);
-          } else if (d.elem_category == 'port_group' && !activePGs.includes(node)) {
-            activePGs.push(node);
-          } else if (node.isEdge() && !activeEdges.includes(node)) {
-            activeEdges.push(node);
-          } else if (d.label == 'group_box' && !activeGBs.includes(node)) {
-            activeGBs.push(node)
-          } else if (d.label == 'map_background' && !activeMBs.includes(node)) {
-            activeMBs.push(node)
-          } else if (d.elem_category == 'map_link' && !activeMapLinks.includes(node)) {
-            activeMapLinks.push(node);
-          }
-        })
+        this.store.dispatch(selectAllElementsOnMap());
         cy.nodes().select();
         cy.edges().select();
-        this.store.dispatch(retrievedMapSelection({ data: true }));
       },
       hasTrailingDivider: false,
       disabled: false,
     }
   }
 
-  getDeselectAllMenu(cy: any, activeNodes: any[], activePGs: any[], activeEdges: any[], activeGBs: any[], activeMBs: any[], activeMapLinks: any[]) {
+  getDeselectAllMenu(cy: any) {
     return {
       id: "deselect_all",
       content: "Deselect All",
@@ -167,15 +147,9 @@ export class CMMapService {
       onClickFunction: (event: any) => {
         cy.nodes().selectify();
         cy.edges().selectify();
-        activeNodes.splice(0);
-        activeGBs.splice(0)
-        activePGs.splice(0);
-        activeEdges.splice(0);
-        activeMBs.splice(0);
-        activeMapLinks.splice(0);
+        this.store.dispatch(unSelectAllElementsOnMap());
         cy.nodes().unselect();
         cy.edges().unselect();
-      this.store.dispatch(retrievedMapSelection({ data: true }));
       },
       hasTrailingDivider: false,
       disabled: false,
